requirements.txt
streamlit
requests
matplotlib
numpy
google-generativeai==0.8.5
textblob